[rocm-libraries] ROCm/rocm-libraries#2848 (commit 5e82c06)
[rocWMMA] Fix standalone Windows build and MinSizeRel build
 on Windows (#2848)

## Motivation

Building rocWMMA as part of TheRock on Windows worked fine, but when
trying to do a standalone build we run into issues with GTest. The
MinSizeRel build was also broken on Windows.

## Technical Details

This change sets a GTest CMake variable to ensure we build the GTest
_static_ library so that it _dynamically_ links against the C runtime if
we are building the tests as dynamic binaries.

Also make sure we don't remove important automatically-added build flags
for the MinSizeRel build on Windows.

## Test Plan

Manual testing on Windows.

## Test Result

Executables linked and ran properly.

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.